backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
yt-dlp==2025.11.12
gtts==2.5.1
openai-whisper
torch
torchaudio
Pillow
requests==2.31.0
pydantic>=2.10.0
pandas
python-dotenv==1.0.0
markdown==3.5.1
langdetect==1.0.9
moviepy==1.0.3
google-genai>=1.0.0
google-api-python-client==2.110.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.1
msgraph-sdk==1.0.0
azure-identity==1.15.0
msal==1.25.0
